requirements.txt
langchain
langchain-community
python-dotenv
langchain-google-genai
streamlit